Refactor MapCanvas into modular components with improved code organization
