+++ src/components/Canvas/components/index.ts
@@ -1,0 +1,5 @@
+export { Territory } from './Territory';
+export { DrawingPreview } from './DrawingPreview';
+export { ZombieArrow } from './ZombieArrow';
+export { ZombieNumber } from './ZombieNumber';
+export { EditorPanel } from './EditorPanel';
