+++ src/components/Canvas/constants/colors.ts
@@ -1,0 +1,31 @@
+export const ZOMBIE_ARROW_ORDER_COLORS = [
+  '#f44336', // Primary
+  '#ff9800', // Secondary
+  '#ffeb3b', // Tertiary
+  '#4caf50', // Quaternary
+  '#00bcd4', // Quinary
+  '#3f51b5', // Senary
+  '#9c27b0', // Septenary
+  '#e040fb', // Octonary
+];
+
+export const ZOMBIE_ARROW_COLORS = [
+  '#f44336', '#ff9800', '#ffeb3b', '#4caf50', '#00bcd4', '#3f51b5', '#9c27b0', '#e040fb', '#222', '#fff'
+];
+
+export const ZOMBIE_COLORS = [
+  '#fff', '#222', '#f44336', '#ff9800', '#ffeb3b', '#4caf50', '#00bcd4', '#3f51b5', '#9c27b0', '#e040fb'
+];
+
+export const ZOMBIE_ARROW_LABELS: { [color: string]: string } = {
+  '#f44336': 'Primary',
+  '#ff9800': 'Secondary',
+  '#ffeb3b': 'Tertiary',
+  '#4caf50': 'Quaternary',
+  '#00bcd4': 'Quinary',
+  '#3f51b5': 'Senary',
+  '#9c27b0': 'Septenary',
+  '#e040fb': 'Octonary',
+  '#222': 'Black',
+  '#fff': 'White',
+};
