+++ src/components/Canvas/constants/fonts.ts
@@ -1,0 +1,13 @@
+export const AVAILABLE_FONTS = [
+  'Arial',
+  'Verdana',
+  'Helvetica',
+  'Times New Roman',
+  'Courier New',
+  'Georgia',
+  'Trebuchet MS',
+  'Impact',
+  'Comic Sans MS'
+];
+
+export const ZOMBIE_FONTS = AVAILABLE_FONTS;
