+++ src/components/Canvas/constants/index.ts
@@ -1,0 +1,3 @@
+export * from './fonts';
+export * from './colors';
+export * from './sizes';
