+++ src/components/Canvas/constants/sizes.ts
@@ -1,0 +1,6 @@
+export const CANVAS_INITIAL_WIDTH = 1200;
+export const CANVAS_INITIAL_HEIGHT = 800;
+
+export const MAX_OFFSET = 18; // px
+export const BASE_OFFSET = 8; // px
+export const OFFSET_STEP = 6; // px
