+++ src/components/Canvas/hooks/index.ts
@@ -1,0 +1,2 @@
+export { usePanZoom } from './usePanZoom';
+export { useDrawing } from './useDrawing';
