+++ src/components/Canvas/utils/index.ts
@@ -1,0 +1,3 @@
+export * from './geometry';
+export * from './canvasUtils';
+export * from './zombieUtils';
